+++ Mariana/costs.py
@@ -14,7 +14,7 @@ class Cost_ABC(Abstraction_ABC) :
         hyps = {}
         for k in self.hyperParameters :
             hyps[k] = getattr(self, k)
-
+        
         message = "%s uses cost %s for %s" % (layer.name, self.__class__.__name__, purpose)
         layer.network.logLayerEvent(layer, message, hyps)
         return self.costFct(targets, outputs)