+++ Mariana/settings.py
@@ -14,3 +14,8 @@ numpy.random.seed(RANDOM_SEED)
 TYPE_INPUT_LAYER = "input"
 TYPE_OUTPUT_LAYER = "output"
 TYPE_HIDDEN_LAYER = "hidden"
+
+import theano
+DEVICE_IS_GPU = (theano.config.device.find("gpu") > -1)
+INTX="int32"
+FLOATX=theano.config.floatX